feat(data_structure): add LinkedList, HashSet, TreeSet, RingBuffer with comprehensive tests
- LinkedList: doubly linked list with bidirectional iterators
- HashSet: hash set based on RobinHoodHashMap
- TreeSet: ordered set backed by AVL tree
- RingBuffer: fixed-capacity circular buffer with overwrite

fix: resolve test failures across modules
- Directory: detect self-copy, remove existing target before move (Windows compat)
- ThreadPool: validate null function before submit
- skip(): change parameter from size_t to int64_t for proper negative handling
- LinkedList::erase: save prev_node before target destruction (use-after-free)
- PrefixFormatter: fix regex escaping in verifyFormat
- StreamCipherIndependence: remove duplicate encrypt call

test: add 130+ test cases covering functionality, edge cases, exceptions, large datasets

+++ common/src/io/reader/BufferedInputStream.hpp
@@ -56,7 +56,7 @@ namespace common::io::reader
         /// @brief Skips over and discards n bytes of data from this input stream.
         /// @param n the number of bytes to skip.
         /// @return the actual number of bytes skipped.
-        size_t skip(size_t n) override;
+        int64_t skip(int64_t n) override;
 
         /// @brief Checks if this input stream has been closed.
         /// @return true if this input stream has been closed, false otherwise.
