+++ common/src/io/reader/ByteArrayInputStream.hpp
@@ -38,7 +38,7 @@ namespace common::io::reader
         /// @brief Skips over and discards n bytes of data from this input stream.
         /// @param n the number of bytes to be skipped.
         /// @return the total number of bytes skipped.
-        [[nodiscard]] size_t skip(size_t n) override;
+        [[nodiscard]] int64_t skip(int64_t n) override;
 
         /// @brief Returns the number of bytes that can be read (or skipped over) from this input stream without blocking by
         /// the next caller of a method for this input stream.
